+++ lib/customize.php
@@ -10,32 +10,6 @@
  * @link    http://www.studiopress.com/
  */
 
-/**
- * Get default link color for Customizer.
- *
- * Abstracted here since at least two functions use it.
- *
- * @since 2.2.3
- *
- * @return string Hex color code for link color.
- */
-function genesis_sample_customizer_get_default_link_color() {
-	return '#c3251d';
-}
-
-/**
- * Get default accent color for Customizer.
- *
- * Abstracted here since at least two functions use it.
- *
- * @since 2.2.3
- *
- * @return string Hex color code for accent color.
- */
-function genesis_sample_customizer_get_default_accent_color() {
-	return '#c3251d';
-}
-
 add_action( 'customize_register', 'genesis_sample_customizer_register' );
 /**
  * Register settings and controls with the Customizer.
@@ -62,9 +36,9 @@ function genesis_sample_customizer_register() {
 			'genesis_sample_link_color',
 			array(
 				'description' => __( 'Change the default color for linked titles, menu links, post info links and more.', 'genesis-sample' ),
-			    'label'       => __( 'Link Color', 'genesis-sample' ),
-			    'section'     => 'colors',
-			    'settings'    => 'genesis_sample_link_color',
+				'label'       => __( 'Link Color', 'genesis-sample' ),
+				'section'     => 'colors',
+				'settings'    => 'genesis_sample_link_color',
 			)
 		)
 	);
@@ -83,9 +57,9 @@ function genesis_sample_customizer_register() {
 			'genesis_sample_accent_color',
 			array(
 				'description' => __( 'Change the default color for button hovers.', 'genesis-sample' ),
-			    'label'       => __( 'Accent Color', 'genesis-sample' ),
-			    'section'     => 'colors',
-			    'settings'    => 'genesis_sample_accent_color',
+				'label'       => __( 'Accent Color', 'genesis-sample' ),
+				'section'     => 'colors',
+				'settings'    => 'genesis_sample_accent_color',
 			)
 		)
 	);