+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Genesis Sample Theme Changelog
 
 ## [Unreleased][unreleased]
-
-_No changes._
+* Add WooCommerce support to the theme and customizer
+* Add an updated and improved responsive menu script
+* Updated the file structure
+* Update Documentation
 
 ## [2.2.4] - 2016-06-08
 * Reorder font size on breadcrumbs